Move MessageRef to protocol.common package
