+++ src/main/java/com/streamr/client/protocol/control_layer/ResendFromRequestAdapter.java
@@ -2,7 +2,7 @@ package com.streamr.client.protocol.control_layer;
 
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.JsonWriter;
-import com.streamr.client.protocol.message_layer.MessageRef;
+import com.streamr.client.protocol.common.MessageRef;
 import com.streamr.client.protocol.message_layer.MessageRefAdapter;
 import java.io.IOException;
