+++ src/main/java/com/streamr/client/protocol/message_layer/StreamMessageV30Adapter.java
@@ -4,6 +4,7 @@ import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.JsonDataException;
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.JsonWriter;
+import com.streamr.client.protocol.common.MessageRef;
 import com.streamr.client.protocol.message_layer.StreamMessage.MessageType;
 import com.streamr.client.protocol.message_layer.StreamMessage.SignatureType;
 import java.io.IOException;
