+++ src/main/java/com/streamr/client/subs/BasicSubscription.java
@@ -4,7 +4,7 @@ import com.streamr.client.MessageHandler;
 import com.streamr.client.exceptions.GapDetectedException;
 import com.streamr.client.exceptions.UnableToDecryptException;
 import com.streamr.client.exceptions.UnsupportedMessageException;
-import com.streamr.client.protocol.message_layer.MessageRef;
+import com.streamr.client.protocol.common.MessageRef;
 import com.streamr.client.protocol.message_layer.StreamMessage;
 import com.streamr.client.utils.Address;
 import com.streamr.client.utils.DecryptionQueues;
