+++ src/test/groovy/com/streamr/client/protocol/message_layer/StreamrSpecification.groovy
@@ -1,5 +1,6 @@
 package com.streamr.client.protocol.message_layer
 
+import com.streamr.client.protocol.common.MessageRef
 import com.streamr.client.utils.Address
 import spock.lang.Specification
