+++ src/test/groovy/com/streamr/client/utils/OrderingUtilSpec.groovy
@@ -1,6 +1,6 @@
 package com.streamr.client.utils
 
-import com.streamr.client.protocol.message_layer.MessageRef
+import com.streamr.client.protocol.common.MessageRef
 import com.streamr.client.protocol.message_layer.StreamMessage
 import com.streamr.client.protocol.message_layer.StreamrSpecification
 import java.util.function.Consumer
