+++ src/test/groovy/com/streamr/client/utils/SigningUtilSpec.groovy
@@ -2,7 +2,7 @@ package com.streamr.client.utils
 
 import com.streamr.client.protocol.message_layer.Json
 import com.streamr.client.protocol.message_layer.MessageId
-import com.streamr.client.protocol.message_layer.MessageRef
+import com.streamr.client.protocol.common.MessageRef
 import com.streamr.client.protocol.message_layer.StreamMessage
 import com.streamr.client.protocol.message_layer.StreamrSpecification
 import org.web3j.crypto.ECKeyPair
