+++ src/test/java/com/streamr/client/testing/StreamMessageExamples.java
@@ -2,7 +2,7 @@ package com.streamr.client.testing;
 
 import com.streamr.client.protocol.message_layer.Json;
 import com.streamr.client.protocol.message_layer.MessageId;
-import com.streamr.client.protocol.message_layer.MessageRef;
+import com.streamr.client.protocol.common.MessageRef;
 import com.streamr.client.protocol.message_layer.StreamMessage;
 import com.streamr.client.utils.Address;
 import java.util.HashMap;
